ZON-5594: Declare UI dependencies separately
