+++ core/setup.py
@@ -18,47 +18,27 @@ setup(
         'BTrees',
         'Jinja2',
         'Pillow',
-        'Pygments',
-        'ZConfig',
         'ZODB',
         'beautifulsoup4',
         'bugsnag',
-        'cssutils',
         'celery >= 4.0',
         'celery_longterm_scheduler',
-        'docutils',
         'elasticsearch >=2.0.0, <3.0.0',
-        'fanstatic',
         'fb',
         'filemagic',
         'gocept.cache >= 2.1',
-        'gocept.fckeditor[fanstatic]',
-        'gocept.form[formlib]>=0.7.5',
         'gocept.lxml>=0.2.1',
-        'gocept.pagelet',
         'gocept.runner>0.5.3',
         'grokcore.component',
-        'guppy',
         'iso8601>=0.1.2',
-        'js.backbone',
-        'js.cropper',
-        'js.handlebars',
-        'js.jquery',
-        'js.jqueryui',
-        'js.mochikit',
-        'js.select2',
-        'js.underscore',
-        'js.vanderlee_colorpicker',
         'lxml>=2.0.2',
         'martian',
         'mock',
         'persistent',
-        'pypandoc',
         'pyramid_dogpile_cache2',
         'pytz',
         'requests',
         'redis',
-        'repoze.vhm',
         'setuptools',
         'sprout',
         'transaction',
@@ -68,93 +48,114 @@ setup(
         'werkzeug',
         'xml-compare',
         'z3c.celery >= 1.2.0.dev0',
-        'z3c.conditionalviews>=1.0b2.dev-r91510',
-        'z3c.flashmessage',
-        'z3c.menu.simple>=0.5.1',
-        'z3c.noop',
-        'z3c.traverser',
-        'zc.datetimewidget',
-        'zc.form',
+        'z3c.traverser',  # XXX Should move to [ui], but is entrenched
         'zc.iso8601',
         'zc.queue',
         'zc.relation',
         'zc.set',
         'zc.sourcefactory',
-        'zc.table',
-        'zeep',
-        'zeit.optivo',
         'zope.annotation',
-        'zope.app.authentication',
-        'zope.app.applicationcontrol',
         'zope.app.appsetup',
-        'zope.app.basicskin',
-        'zope.app.broken',
-        'zope.app.component>=3.4.0b3',
-        'zope.app.container',
-        'zope.app.content',
-        'zope.app.dependable',
-        'zope.app.error',
-        'zope.app.exception',
         'zope.app.file',
-        'zope.app.folder',
-        'zope.app.form>=3.6.0',
-        'zope.app.generations',
-        'zope.app.interface',
-        'zope.app.http',
         'zope.app.keyreference',
-        'zope.app.localpermission',
         'zope.app.locking',
-        'zope.app.pagetemplate',
-        'zope.app.preference',
-        'zope.app.principalannotation',
-        'zope.app.publication',
-        'zope.app.publisher',
-        'zope.app.renderer',
-        'zope.app.rotterdam',
-        'zope.app.schema',
         'zope.app.security',
-        'zope.app.securitypolicy',
-        'zope.app.tree',
-        'zope.app.wsgi',
-        'zope.app.zapi',  # undeclared by z3c.menu.simple
-        'zope.app.zopeappgenerations',
-        'zope.authentication',
-        'zope.browser',
-        'zope.browserpage',
         'zope.cachedescriptors',
-        'zope.configuration',
         'zope.component',
         'zope.container>=3.8.1',
         'zope.copypastemove',
-        'zope.deferredimport',  # undeclared by zc.form
         'zope.dottedname',
         'zope.dublincore',
-        'zope.error',
         'zope.event',
         'zope.exceptions',
         'zope.file',
-        'zope.formlib',
         'zope.i18n>3.4.0',
         'zope.index',
         'zope.interface',
         'zope.lifecycleevent',
         'zope.location>=3.4.0b2',
-        'zope.login',
-        'zope.password',
-        'zope.pluggableauth',
-        'zope.principalannotation',
         'zope.proxy',
-        'zope.publisher',
+        'zope.publisher',  # XXX Should move to [ui], but is entrenched
         'zope.schema',
         'zope.security',
         'zope.securitypolicy',
-        'zope.session',
-        'zope.site',
-        'zope.traversing',
-        'zope.viewlet',
+        'zope.traversing',  # XXX Should move to [ui], but is entrenched
         'zope.xmlpickle',
     ],
     extras_require={
+        'ui': [
+            'Pygments',
+            'ZConfig',
+            'cssutils',
+            'docutils',
+            'fanstatic',
+            'gocept.fckeditor[fanstatic]',
+            'gocept.form[formlib]>=0.7.5',
+            'gocept.pagelet',
+            'guppy',
+            'js.backbone',
+            'js.cropper',
+            'js.handlebars',
+            'js.jquery',
+            'js.jqueryui',
+            'js.mochikit',
+            'js.select2',
+            'js.underscore',
+            'js.vanderlee_colorpicker',
+            'pypandoc',
+            'repoze.vhm',
+            'z3c.conditionalviews>=1.0b2.dev-r91510',
+            'z3c.flashmessage',
+            'z3c.menu.simple>=0.5.1',
+            'z3c.noop',
+            'zc.datetimewidget',
+            'zc.form',
+            'zope.deferredimport',  # undeclared by zc.form
+            'zc.table',
+            'zeep',
+            'zeit.optivo',
+            'zope.app.authentication',
+            'zope.app.applicationcontrol',
+            'zope.app.basicskin',
+            'zope.app.broken',
+            'zope.app.component>=3.4.0b3',
+            'zope.app.container',
+            'zope.app.content',
+            'zope.app.dependable',
+            'zope.app.error',
+            'zope.app.exception',
+            'zope.app.folder',
+            'zope.app.form>=3.6.0',
+            'zope.app.generations',
+            'zope.app.interface',
+            'zope.app.http',
+            'zope.app.localpermission',
+            'zope.app.pagetemplate',
+            'zope.app.preference',
+            'zope.app.principalannotation',
+            'zope.app.publication',
+            'zope.app.publisher',
+            'zope.app.renderer',
+            'zope.app.rotterdam',
+            'zope.app.schema',
+            'zope.app.securitypolicy',
+            'zope.app.tree',
+            'zope.app.wsgi',
+            'zope.app.zapi',  # undeclared by z3c.menu.simple
+            'zope.app.zopeappgenerations',
+            'zope.authentication',
+            'zope.browser',
+            'zope.browserpage',
+            'zope.configuration',
+            'zope.error',
+            'zope.formlib',
+            'zope.login',
+            'zope.pluggableauth',
+            'zope.principalannotation',
+            'zope.session',
+            'zope.site',
+            'zope.viewlet',
+        ],
         'test': [
             'cssselect',
             'docker',
