+++ core/src/zeit/cms/tagging/browser/configure.zcml
@@ -45,4 +45,16 @@
     permission="zope.View"
     />
 
+  <!-- XXX AbsoluteURL can be accessed in two ways, and one doesn't delegate to
+       the other -->
+  <adapter
+      factory=".tag.AbsoluteURL"
+      provides="zope.traversing.browser.interfaces.IAbsoluteURL"
+      />
+  <adapter
+      factory=".tag.AbsoluteURL"
+      provides="zope.interface.Interface"
+      name="absolute_url"
+      />
+
 </configure>
